// Observations-all/ready_mobile_farm_proj/mobile_device_farm/src/app/models/device.ts
export class Device {
  serialNo: string;
  name: string;
  version: string;
  os?: string;
  isAvailable?: string;
  platform: string;
  id: string;
  userId: string;
  osVersion: string;
  deviceStatus: string;
  model: string;
  buildNo: string;
  sdkVersion: string;
  storage: string;
  ram: string;
  remarks: string;
  ipAddress: string;
  hardware: string;
  hardware_version: string;
  height: string;
  width: string;
  createdDate: string;
  status?: string;
  isInUse?: string;

  updatedDate;
}